preferentially choose high naive-scoring words
This was probably worth trying, but still isn't close to right. The
idea was that, e.g., if you're placing a word in the last row, you
should prefer words with letters that are frequently final letters, in
the hopes of having better choices for the columns. But not only have
I not yet gotten around to disallowing the formation of nonwords in
the other direction, the naive scoring method isn't really
appropriate, either (e.g., way too likely to place double-Ss).