this still doesn't work
I thought that this would prevent the program from forming nonwords in
the other direction while placing words, by checking the constraints
in the other direction. But I'm confused and it doesn't
work. Apparently I think it's okay to commit broken code when you're
still in the exploratory scratch phase.