second constraints
It looks like my struggles yesterday were due to trying to use the
return value of "substitute", which, in its previous incarnation,
returned the keyword it swapped in to the relevant atom, rather than
the sequence of atoms I've been calling a line. In this commit,
"overwrite" and "substitute" get bangs as a reminder of their status
as mutators and return the line they act on.

Incidentally, did you know that interactive staging is nice when
you're working on your next thing but realize that you forgot to do
something that logically belongs in the previous commit?